# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DuitPlan API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /debts:
    get:
      operationId: listDebts
      tags: [debts]
      summary: List user debts
      responses:
        "200":
          description: List of debts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debt"

    post:
      operationId: createDebt
      tags: [debts]
      summary: Create a debt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDebtBody"
      responses:
        "201":
          description: Debt created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
  /debts/{id}:
    patch:
      operationId: updateDebt
      tags: [debts]
      summary: Update a debt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDebtBody"
      responses:
        "200":
          description: Updated debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"

    delete:
      operationId: deleteDebt
      tags: [debts]
      summary: Delete a debt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /debts/{id}/simulate:
    post:
      operationId: simulateDebtPayoff
      tags: [debts]
      summary: Simulate debt payoff with extra monthly payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebtSimulateBody"
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtScenario"

  # Insights
  /insights:
    get:
      operationId: listInsights
      tags: [insights]
      summary: Get financial insights for a month
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
          description: "Month in YYYY-MM format"
      responses:
        "200":
          description: List of insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insight"
  /insights/generate:
    post:
      operationId: generateInsights
      tags: [insights]
      summary: Generate/refresh insights for a month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInsightsBody"
      responses:
        "200":
          description: Generated insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insight"

  # Dashboard
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get monthly dashboard summary
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
          description: "Month in YYYY-MM format"
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/spending-by-category:
    get:
      operationId: getSpendingByCategory
      tags: [dashboard]
      summary: Get spending breakdown by category for a month
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
          description: "Month in YYYY-MM format"
      responses:
        "200":
          description: Spending by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySpend"
  /dashboard/recent-transactions:
    get:
      operationId: getRecentTransactions
      tags: [dashboard]
      summary: Get recent transactions for dashboard
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  # Uploads
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        accountId:
          type: ["string", "null"]
        categoryId:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        date:
          type: string
        amount:
          type: string
        type:
          type: string
        description:
          type: string
        merchant:
          type: ["string", "null"]
        source:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - date
        - amount
        - type
        - description
        - source
        - createdAt
        - updatedAt
...
    Debt:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        debtType:
          type: string
        lender:
          type: string
        outstandingBalance:
          type: string
        monthlyPayment:
          type: string
        interestRate:
          type: ["string", "null"]
        targetExtraPayment:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - debtType
        - lender
        - outstandingBalance
        - monthlyPayment
        - createdAt
        - updatedAt
    CreateDebtBody:
      type: object
      properties:
        debtType:
          type: string
        lender:
          type: string
        outstandingBalance:
          type: string
        monthlyPayment:
          type: string
        interestRate:
          type: ["string", "null"]
      required:
        - debtType
        - lender
        - outstandingBalance
        - monthlyPayment
    UpdateDebtBody:
      type: object
      properties:
        debtType:
          type: ["string", "null"]
        lender:
          type: ["string", "null"]
        outstandingBalance:
          type: ["string", "null"]
        monthlyPayment:
          type: ["string", "null"]
        interestRate:
          type: ["string", "null"]
        targetExtraPayment:
          type: ["string", "null"]
    DebtSimulateBody:
      type: object
      properties:
        extraMonthlyPayment:
          type: string
      required:
        - extraMonthlyPayment
    DebtScenario:
      type: object
      properties:
        id:
          type: string
        debtId:
          type: string
        extraMonthlyPayment:
          type: string
        estimatedMonthsSaved:
          type: ["integer", "null"]
        estimatedPayoffDate:
          type: ["string", "null"]
        basePayoffMonths:
          type: ["integer", "null"]
        newPayoffMonths:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - debtId
        - extraMonthlyPayment
        - createdAt

    # Insights
    Insight:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        month:
          type: string
        title:
          type: string
        message:
          type: string
        severity:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - month
        - title
        - message
        - severity
        - createdAt
    GenerateInsightsBody:
      type: object
      properties:
        month:
          type: string
      required:
        - month

    # Dashboard
    DashboardSummary:
      type: object
      properties:
        month:
          type: string
        monthlyIncome:
          type: string
        totalCommitments:
          type: string
        totalSpent:
          type: string
        remaining:
          type: string
        totalDebtMonthlyPayment:
          type: string
        debtToIncomeRatio:
          type: string
        transactionCount:
          type: integer
      required:
        - month
        - monthlyIncome
        - totalCommitments
        - totalSpent
        - remaining
        - totalDebtMonthlyPayment
        - debtToIncomeRatio
        - transactionCount
    CategorySpend:
      type: object
      properties:
        categoryId:
          type: ["string", "null"]
        categoryName:
          type: string
        totalSpent:
          type: string
        percentage:
          type: number
      required:
        - categoryName
        - totalSpent
        - percentage

    # Upload
...