# lib/api-zod/src/generated/types/debtScenario.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DuitPlan API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DebtScenario {
  id: string;
  debtId: string;
  extraMonthlyPayment: string;
  /** @nullable */
  estimatedMonthsSaved?: number | null;
  /** @nullable */
  estimatedPayoffDate?: string | null;
  /** @nullable */
  basePayoffMonths?: number | null;
  /** @nullable */
  newPayoffMonths?: number | null;
  createdAt: string;
}